refactor prodject so that file structure is in line with the Jekyll plugin format
