+++ lib/jekyll-notion/version.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Jekyll
+  module Notion
+    VERSION = "0.1.0"
+  end
+end
