--- lib/notion_to_jekyll/version.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module NotionToJekyll
-  VERSION = "0.1.0"
-end
