+++ spec/fixtures/posts/post.md
@@ -1,0 +1,18 @@
+## I’m a Heading
+Here’s a list of things I like to do:
+- Foo
+- Bar
+- Baz
+
+Some Code:
+{% highlight ruby %}
+beans.each do |bean|
+	bean.compute! if bean.computable?
+end
+{% endhighlight %}
+
+Here’s a screen shot:
+![a red ruby](assets/images/a_red_ruby.png)
+### In Conclusion
+I can’t think of anything else to say
+
